+++ sms_frame_mail_campaign/__openerp__.py
@@ -1,6 +1,6 @@
 {
     'name': "SMS Framework - SMS Campaign",
-    'version': "1.0.1",
+    'version': "1.0.2",
     'author': "Sythil Tech",
     'category': "Tools",
     'summary': "Adds the ability to more easily send smses from a marketing campaign",