+++ sms_frame/doc/changelog.rst
@@ -1,3 +1,7 @@
+v1.0.8
+======
+* SMS queuing in position for mass sms
+
 v1.0.7
 ======
 * MMS template sending fix