+++ sms_frame_mail_campaign/doc/changelog.rst
@@ -1,3 +1,7 @@
+v1.0.2
+======
+* Just queue the smses
+
 v1.0.1
 ======
 * MMS Fix